+++ dao-web-app/src/components/Dao/DaoDiscussionPage.tsx
@@ -2,18 +2,39 @@ import { IDAOState } from "@daostack/client";
 import { DiscussionEmbed } from "disqus-react";
 import * as React from "react";
 import { BreadcrumbsItem } from "react-breadcrumbs-dynamic";
-import * as Sticky from "react-stickynode";
+// import * as Sticky from "react-stickynode";
 import * as css from "./Dao.scss";
+import Box = require("3box");
+import { IProfileState } from "reducers/profilesReducer";
+// import { getWeb3Provider } from "arc";
 
 import moment = require("moment");
 
 interface IProps {
   dao: IDAOState;
+  currentAccountAddress: string;
+  currentAccountProfile: IProfileState;
 }
 
 export default class DaoDiscussionPage extends React.Component<IProps, null> {
 
   public async componentDidMount() {
+    console.log("Discuss props ", this.props);
+    const provider = await  Box.get3idConnectProvider(); // recomended provider
+    const box = await Box.openBox(this.props.currentAccountAddress, provider)
+    const space = await box.openSpace('snglsGeneral')
+    const thread = await space.joinThread('myThread', {
+      firstModerator: "0x4699f05bF35Ba6820F8393C44F780fBe00fe7aa9",
+      members: true
+    });
+    await thread.addMember(this.props.currentAccountAddress);
+
+    await thread.post('hello world')
+
+    console.log("Thread posts: ", await thread.getPosts());
+
+    await space.syncDone
+
     localStorage.setItem(`daoWallEntryDate_${this.props.dao.address}`, moment().toISOString());
   }
 
@@ -30,11 +51,11 @@ export default class DaoDiscussionPage extends React.Component<IProps, null> {
       <div>
         <BreadcrumbsItem to={"/dao/" + dao.address + "/discussion"}>Discussion</BreadcrumbsItem>
 
-        <Sticky enabled top={50} innerZ={10000}>
+        {/* <Sticky enabled top={50} innerZ={10000}> */}
           <div className={css.daoHistoryHeader}>
             Discuss {dao.name}
           </div>
-        </Sticky>
+        {/* </Sticky> */}
 
         <div>
           <DiscussionEmbed shortname={process.env.DISQUS_SITE} config={disqusConfig}/>